fix: remove debug file
