+++ packages/api/src/routes/nfts-upload.js
@@ -2,12 +2,10 @@ import { HTTPError } from '../errors.js'
 import * as cluster from '../cluster.js'
 import { JSONResponse } from '../utils/json-response.js'
 import { validate } from '../utils/auth.js'
-import { debug } from '../utils/debug.js'
 import { toNFTResponse } from '../utils/db-transforms.js'
 import { parseCid } from '../utils/utils.js'
 import * as constants from '../constants.js'
 
-const log = debug('nfts-upload')
 const LOCAL_ADD_THRESHOLD = constants.cluster.localAddThreshold
 
 /**
